support shows and episodes authorization

+++ src/actions/DataActions.js
@@ -17,27 +17,39 @@ const getDataActions = (dataType) => ({
 		data,
 	}),
 
-	changeStatus: (id, status) => ({
-		type: ACTIONS.CHANGE_STATUS,
-		payload: http.put(`/${dataType}/`, {
-			[dataType + "_id"]: id,
-			status: status,
-		}),
-		meta: {
-			id,
-			status: mapStatus.get(status),
-			dataType,
-		},
-	}),
+	changeStatus: (id, status) => {
+		let payload;
+		if (id.length) {
+			payload = http.put(`/${dataType}/`, {
+				[dataType + "_id"]: id,
+				status: status,
+			});
+		} else {
+			payload = Promise.resolve([]);
+		}
+
+		return {
+			type: ACTIONS.CHANGE_STATUS,
+			payload,
+			meta: {
+				id,
+				status: mapStatus.get(status),
+				dataType,
+			},
+		};
+	},
 
 	deleteData: (id) => {
 		let payload;
 		if (id instanceof Array) {
-			payload = http.delete(`/${dataType}/`, {
-				data: {
-					["deleted_" + dataType + "_id"]: id,
-				},
-			});
+			if (id.length === 0) payload = Promise.resolve([]);
+			else {
+				payload = http.delete(`/${dataType}/`, {
+					data: {
+						["deleted_" + dataType + "_id"]: id,
+					},
+				});
+			}
 		} else {
 			payload = http.delete(`/${dataType}/${id}`);
 		}
